--- a/torch/optim/_muon.py
+++ b/torch/optim/_muon.py
@@ -302,11 +302,8 @@ def _single_tensor_muon(
             raise ValueError("Param gradient must be a 2D matrix")
 
         buf = muon_momentum_bufs[i]
-        buf.mul_(momentum).add_(grad)
-        if nesterov:
-            grad = grad.add(buf, alpha=momentum)
-        else:
-            grad = buf
+        buf.lerp_(grad, 1 - momentum)
+        grad = grad.lerp_(buf, momentum) if nesterov else buf
 
         update = _zeropower_via_newtonschulz(grad, ns_coefficients, ns_steps, eps)