--- a/torch/optim/_muon.py
+++ b/torch/optim/_muon.py
@@ -303,9 +303,9 @@ def _single_tensor_muon(
 
         buf = muon_momentum_bufs[i]
         buf.lerp_(grad, 1 - momentum)
-        grad = grad.lerp(buf, momentum) if nesterov else buf
+        update = grad.lerp(buf, momentum) if nesterov else buf
 
-        update = _zeropower_via_newtonschulz(grad, ns_coefficients, ns_steps, eps)
+        update = _zeropower_via_newtonschulz(update, ns_coefficients, ns_steps, eps)
 
         adjusted_lr = _adjust_lr(lr, adjust_lr_fn, param.shape)