… docs/source/optim.md
… test/inductor/test_compiled_optimizers.py
… test/test_optim.py