@janeyx99 2025-08-12
Given that we had agreed to land the simplest single device Muon into torch/optim as our first step, it'd be clearest to land what people accept as the original implementation as defined in Keller Jordan's blog ( https://kellerjordan.github.io/posts/muon/). As this implementation chooses newton schulz as the algo, we should take the same stance. This means we can simplify the constructor API greatly (I will get to extensibility right after):
- Remove the msign_fn callable argument (the algorithm will just call NS by default today). OSS folks will not expect to pass in a callable to the constructor, so we will not accept the PR with a kwarg that takes in a callable. Instead, for customization, we can intake a string enum (more on this later).
- Remove the struct definitions and flatten the kwargs as top-level keyword arguments to the constructor. Having layers of configs is confusing and unnecessarily abstracted. Since we are going in on NS being the algorithm for single device Muon, we can explicitly list out these kwargs in the constructor.
- Move the algorithm description of NS into the Muon doc, so people can see immediately what algo they are calling.
- I am remembering that the original test scripts comparing this implementation to Keller Jordan's had high atol and rtols, which is surprising as the two algorithms both use PyTorch ops in python, and so I'd expect the results to be the same. Could you link a standalone script that can be run to ascertain correctness and explain why the high atol/rtols are necessary (if they still are)? As we want to land a trustworthy impl for folks to try out, we need to ensure the accuracy results are expected.

I'm realizing that you have interest in extending the algorithm to be distributed (vs another orthogonalization algo for single-device). We are strict on keeping `torch/optim` code single-device runnable and maximally composable, so we cannot land anything distributed in `torch/optim` and I'd propose landing the distributed optimizer solution in `https://github.com/pytorch/pytorch/tree/main/torch/distributed/optim`. With that, I see the possible extension options as below:
a) If we are interested in other orthogonalization techniques for single device, we'd recommend using string enum kwargs similar to `line_search_fn` in https://docs.pytorch.org/docs/stable/generated/torch.optim.LBFGS.html#lbfgs, where the default None is NS, and other strings can represent other algorithms.
b) If we are attempting to extend in a distributed manner, and the code (state_dict, etc) is easily shareable, we'd recommend subclassing Muon into a new distributed optimizer in `torch/distributed/optim`. If the code ends up not being so shareable, it is perfectly acceptable to have Dion or a different optim class entirely living in `torch/distributed/optim`.