@janeyx99 2025-08-19
> updated in the _Numerics_ section!

Thanks! If I understand correctly, if the second dimension is all the same then there should be no significant difference with adjust_lr, right? 5% relative difference is pretty high still. The script uses a Linear(10, 10)--which I thought would qualify for the second dim being the same, and thus I would expect a smaller relative difference. I can expect some difference arising from the math being different (using clamp) but not that much.