@samsja 2025-08-19
> Yes, we want to support something like this more generically too, like an OptimizerDict that would dispatch into the proper optimizers but have one call to step and one unified state dict. There is a next step plan to design this. For now, we are intentionally keeping this PR clean and simple to understand as a standalone piece

yes makes sense, looking forward to it !