@chuanhaozhuge 2025-08-21
> > ( I hope its find for me to add comments here, I am heavily using muon with pytorch and really like the idea of having an official torch implementation)
> 
> Totally good! @chuanhaozhuge I do like the idea of doing the math on the last two dimensions to widen the use case to accept more than 2D params. What do you think?
> 
> > I think that user would also like use a version of the optimizer that can do both Muon and Adam depending on the group. In the official implementation frm keller jordan there is such class that I think most people use https://github.com/KellerJordan/Muon/blob/f90a42b28e00b8d9d2d05865fe90d9f39abcbcbd/muon.py#L138
> 
> Yes, we want to support something like this more generically too, like an OptimizerDict that would dispatch into the proper optimizers but have one call to step and one unified state dict. There is a next step plan to design this. For now, we are intentionally keeping this PR clean and simple to understand as a standalone piece.
> 
> > Also I am wondering if there is a plan to make this muon implementation already efficient with torch DDP ? In contrast with an adam update that can be done on each gpu without too much cost it could be pretty wastefull to do NS on all gradient on every gpu
> 
> Making this implementation distributed is a next step that's been put in motion. Do you have a proposal/use case for this? If so, can you follow up on the issue!

thanks for the input! discussed offline on the choice of restricting to accept 2D params only. updated the PR description.