@janeyx99 2025-08-21
I think it is okay to have the 2D limitation for now, to force users to massage their params according to their desires, instead of always taking the last 2 dims. In any case, supporting more dims can come in a later PR. Let's make sure we document that loudly.

I am also happy with the KellerJordan adjust_lr_fn as default, with optionality for Moonshot's.

My last remaining blocker is that I would expect the numerics to be almost identical (not bitwise but close enough) now that the adjust lr default is keller jordan's, but running the script locally with this change has not been the case:
```
import muon
from muon import SingleDeviceMuon
import torch
import torch.optim
from torch.nn import Linear, MSELoss
import copy

# dummy model and data
model0 = Linear(10, 10, bias=False)
model1 = copy.deepcopy(model0)
inputs = torch.randn(8, 10)
targets = torch.randn(8, 10)
loss = MSELoss()

lr = 1e-3
wd = 0.1
momentum = 0.95

opt_ref_muon = SingleDeviceMuon(
    params=model0.parameters(), 
    lr=lr,
    weight_decay=wd,
    momentum=momentum,
)

opt_exp_muon = torch.optim.Muon(
    params=model1.parameters(),
    lr=lr,
    weight_decay=wd,
    momentum=momentum,
)

out_ref = model0(inputs)
loss_ref = loss(out_ref, targets)
opt_ref_muon.zero_grad()
loss_ref.backward()
opt_ref_muon.step()

out_exp = model1(inputs)
loss_exp = loss(out_exp, targets)
opt_exp_muon.zero_grad()
loss_exp.backward()
opt_exp_muon.step()

for p_ref, p_exp in zip(model0.parameters(), model1.parameters()):
    torch.testing.assert_close(p_ref, p_exp)
```

Returns:
```
(pytorch-3.13) [janeyx@blah /data/users/janeyx/workspace/pytorch (muon)]$ python playground.py 
Traceback (most recent call last):
  File "/data/users/janeyx/workspace/pytorch/playground.py", line 65, in <module>
    torch.testing.assert_close(p_ref, p_exp)
    ~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^
  File "/data/users/janeyx/workspace/pytorch/torch/testing/_comparison.py", line 1589, in assert_close
    raise error_metas[0].to_error(msg)
AssertionError: Tensor-likes are not close!

Mismatched elements: 77 / 100 (77.0%)
Greatest absolute difference: 0.00015881657600402832 at index (8, 5) (up to 1e-05 allowed)
Greatest relative difference: 0.025237105786800385 at index (9, 1) (up to 1.3e-06 allowed)
```